requirements.txt
# To install these Python dependencies, please type:
# pip install -r requirements.txt
gensim==3.8.3
numpy==1.18.4
pandas==1.0.4
scikit-learn==0.23.1
sklearn==0.0
streamlit==0.60.0
plotly==4.8.1
nltk==3.5
sentence-transformers==0.2.6.1
google-cloud-storage==1.29.0
